Remove backslashes from osm ways, as per:
Some of the road/way names in the OSM data seem to contain a backslash ("\") which
they should not. This case is not tested and fixed by the "osm2pgrouting" tool so far
and this most probably causes the error you get. What you could do is inserting the
following line in the in the "exportWays" method in the "Export2DB.cpp" file (in the
"src" folder of your "osm2pgrouting" folder) and after that build the tool again
("make"):

http://gis.stackexchange.com/questions/12070/error-during-importing-a-map-with-osm2pgrouting